Move analytics into own module
