+++ src/core/Analytics/__mocks__/index.tsx
