+++ src/core/Analytics/prepareStepsConfig.tsx
